refactor(agents): share runtime tool quarantine helpers
